requirements.txt
chargebee>=3.9.0
python-dotenv>=1.0.0
rich>=13.0.0
aiohttp>=3.9.0

# Churn analysis dependencies (optional, for analyze command)
matplotlib>=3.7.0
seaborn>=0.12.0
